Adding Titles
The addition of any kind of titles is finally possible. I had started wokring on udpating titles, but then I deleted it once I realized it was not part of the requirement.

I have already started the next column of assingment which is check point.

check point Viiew is already up and running. Tomorrow I will be working on
Check points controller and query .
Rentals