Correction on Rental and Commenting
I have corrected the validation for the title that the customer is trying to purchase. For some reason it didn`t show up the message and I realized it was because the brackets were being closed after the else. So I changed the order in which the validation occurs and now problem is sorted.

I have also focused on commenting and indenting classes