Commenting / Validation
I have finalized working on Commenting and Validation.

Total number of titles rented was not working because my query was returning 0. Usually Sql ignores null rows and in this case I needed to count the null cell because the title had not been returned yet. So in order to make it work I had to change the signal = (equals) for is in my query. Then it worked.